refactor(memory): drop boundary extraction + memoryd daemon
Memory is now captured exclusively via the in-session `remember` tool
(driven by the per-turn nudge, #132) and folded into the summary by the
startup consolidation pass. The boundary-extraction path re-fed an entire
finished session's history to the MAIN model on the next startup to
re-mine durable facts — work the live `remember` calls already did, with
no semantic dedup, so it produced duplicate entries and burned tokens for
no gain. With live capture now reliable, the post-session re-scan earns
nothing.

Removed:
- Boundary extraction: ExtractMemory + extractSystem prompt,
  extractPreviousSession (startup) and extractIdleSession (daemon).
- The `octo memoryd` daemon entirely (cmd + internal/memoryd package,
  start/stop/status, PID file, completion + help wiring). Its sole job
  was async extraction; consolidation already runs at chat startup.
- Per-rollout summaries (only the extraction path produced them):
  SaveRolloutSummary / ListRolloutSummaries / RolloutSummary /
  rollout_summaries dir / State.LastExtractedSession.

Kept: ConsolidateMemory + shared JSON helpers (were cohabiting in
extract.go) split into consolidate.go + jsonparse.go. Consolidation
(consolidateIfDue) still runs at startup once >=5 entries accumulate.

Docs: roadmap + hindsight notes updated to current state; the C9 design
doc gets a superseded banner + per-section annotations, kept as a
historical design record.

make vet / go test -race ./... / gofmt all clean. Net -2062 lines.
