+++ cmd/octo/main_test.go
@@ -76,7 +76,6 @@ func TestRun_HelpWithSubcommand_PrintsRichHelp(t *testing.T) {
 		{"task", []string{"octo task", "Examples:", "ID shortcuts", "octo task start"}},
 		{"memory", []string{"octo memory", "octo memory list"}},
 		{"init", []string{"octo init", ".octorules"}},
-		{"memoryd", []string{"octo memoryd", "PID file", "octo memoryd start"}},
 		{"mcp", []string{"octo mcp", "mcp.json", "mcp__"}},
 		{"completion", []string{"octo completion", "shell-completion"}},
 	}
