+++ cmd/octo/memory_extract.go
@@ -2,8 +2,6 @@ package main
 
 import (
 	"context"
-	"fmt"
-	"io"
 	"strings"
 	"time"
 
@@ -12,75 +10,27 @@ import (
 	"github.com/Leihb/octo-agent/internal/tools"
 )
 
-// Consolidation triggers (Phase 1 fallback; the daemon will own these later).
+// Consolidation trigger thresholds. Memory is captured live via the `remember`
+// tool; this folds those accumulated entries into the summary when due.
 const (
 	consolidateMinEntries = 5
 	consolidateMinAge     = 24 * time.Hour
 )
 
-// maybeProcessMemory runs the boundary-extraction + consolidation triggers at
-// session start (the no-daemon fallback path). It extracts durable facts from
-// the most-recent finished session (skipping excludeID, the one we're about to
-// resume), then consolidates if due. Best-effort: any side-call error is
-// swallowed so a memory hiccup never blocks the session.
-func maybeProcessMemory(a *agent.Agent, store *memory.Store, out io.Writer, excludeID string) {
+// maybeProcessMemory runs the consolidation trigger at session start. It folds
+// the live-captured entries (from the `remember` tool) into the summary when
+// due. Best-effort: any side-call error is swallowed so a memory hiccup never
+// blocks the session.
+func maybeProcessMemory(a *agent.Agent, store *memory.Store) {
 	if store == nil {
 		return
 	}
 	ctx := context.Background()
 	st := store.LoadState()
 
-	extractPreviousSession(ctx, a, store, out, excludeID, &st)
 	consolidateIfDue(ctx, a, store, &st)
 
 	_ = store.SaveState(st)
-}
-
-func extractPreviousSession(ctx context.Context, a *agent.Agent, store *memory.Store, out io.Writer, excludeID string, st *memory.State) {
-	sessions, err := agent.ListSessions(2)
-	if err != nil {
-		return
-	}
-	for _, sess := range sessions {
-		if sess.ID == excludeID {
-			continue // the session we're resuming — still ongoing
-		}
-		if sess.ID == st.LastExtractedSession || sess.TurnCount() == 0 {
-			return // reached the already-extracted boundary (or nothing to do)
-		}
-		res, err := a.ExtractMemory(ctx, sess.ToHistory().Snapshot())
-		if err == nil {
-			n := 0
-			for _, f := range res.Facts {
-				if serr := store.Save(memory.Entry{
-					Description: f.Description,
-					Type:        memory.Type(f.Type),
-					Body:        f.Content,
-				}); serr == nil {
-					n++
-				}
-			}
-			// Persist the per-rollout narrative reference. Fall back to the
-			// session id as the slug when the model returned "" (no-op
-			// extractions return empty here and SaveRolloutSummary skips them
-			// anyway).
-			slug := res.Slug
-			if slug == "" {
-				slug = sess.ID
-			}
-			_ = store.SaveRolloutSummary(slug, res.Summary)
-
-			if n > 0 || res.Summary != "" {
-				suffix := ""
-				if res.Summary != "" {
-					suffix = " (+ rollout summary)"
-				}
-				fmt.Fprintf(out, "  ⓘ remembered %d fact(s) from your last session%s\n", n, suffix)
-			}
-		}
-		st.LastExtractedSession = sess.ID
-		return // only the most-recent session, to bound startup cost
-	}
 }
 
 func consolidateIfDue(ctx context.Context, a *agent.Agent, store *memory.Store, st *memory.State) {
@@ -96,9 +46,9 @@ func consolidateIfDue(ctx context.Context, a *agent.Agent, store *memory.Store, 
 	// Prefer the sub-agent path (#6) when a Spawner is registered: a child
 	// agent runs in its own context with read-only filesystem tools, can
 	// grep / read_file across the memory dir to look at the actual entry
-	// files and rollout summaries, and returns the new summary text. Falls
-	// back to the side-call (priorSummary + newNotes only, no tools) when no
-	// Spawner is wired or the sub-agent declines.
+	// files, and returns the new summary text. Falls back to the side-call
+	// (priorSummary + newNotes only, no tools) when no Spawner is wired or
+	// the sub-agent declines.
 	summary := consolidateViaSubAgent(ctx, store, priorSummary, newNotes)
 	if summary == "" {
 		summary, err = a.ConsolidateMemory(ctx, priorSummary, newNotes)
@@ -140,9 +90,9 @@ var consolidationToolAllowlist = []string{"read_file", "grep", "glob"}
 // returned nothing. The empty case is non-fatal — caller falls back to the
 // side-call.
 //
-// The sub-agent's win over the side-call: it can grep through rollout_summaries
-// for context that didn't make it into newNotes, read the actual <slug>.md
-// frontmatter, and check MEMORY.md for cross-references — autonomously, in its
+// The sub-agent's win over the side-call: it can read the actual <slug>.md
+// frontmatter for context that didn't make it into newNotes, and check
+// MEMORY.md for cross-references — autonomously, in its
 // own context window.
 func consolidateViaSubAgent(ctx context.Context, store *memory.Store, priorSummary, newNotes string) string {
 	spawner := tools.ActiveSpawner()
@@ -184,7 +134,6 @@ func buildConsolidationPrompt(memDir, priorSummary, newNotes string) string {
 		b.WriteString("- Root: " + memDir + "\n")
 		b.WriteString("- " + memDir + "/MEMORY.md (index of slugs)\n")
 		b.WriteString("- " + memDir + "/<slug>.md (one fact per file, with frontmatter)\n")
-		b.WriteString("- " + memDir + "/rollout_summaries/<timestamp>-<slug>.md (per-session narratives)\n")
 		b.WriteString("- " + memDir + "/memory_summary.md (the file you're updating; current contents below)\n")
 	}
 	b.WriteString("\nCurrent consolidated summary (may be empty on first pass):\n\n")
@@ -195,7 +144,7 @@ func buildConsolidationPrompt(memDir, priorSummary, newNotes string) string {
 	}
 	b.WriteString("\n\nNew memory entries since the last consolidation (the index digest):\n\n")
 	b.WriteString(newNotes)
-	b.WriteString("\n\nYour task: produce the UPDATED consolidated summary. Fold the new entries into the current summary, dedupe, drop anything stale or trivial, and keep load-bearing facts. Be terse — bullet points, grouped loosely by kind (who the user is, how they like to work, ongoing project context, useful references). If you need more context than the digest gives you (a specific quote, the rationale behind a feedback fact, a related rollout summary), use read_file / grep / glob to look at the actual files.\n\n")
+	b.WriteString("\n\nYour task: produce the UPDATED consolidated summary. Fold the new entries into the current summary, dedupe, drop anything stale or trivial, and keep load-bearing facts. Be terse — bullet points, grouped loosely by kind (who the user is, how they like to work, ongoing project context, useful references). If you need more context than the digest gives you (a specific quote, the rationale behind a feedback fact), use read_file / grep / glob to look at the actual files.\n\n")
 	b.WriteString("Output ONLY the new summary text. No preamble, no code fences, no commentary about what you changed. The parent will write your output to memory_summary.md (it adds the protocol marker automatically — you don't need to).\n")
 	return b.String()
 }
